refactor: Migrate to Typescript + clean project structure
