+++ package.json
@@ -11,22 +11,30 @@
   },
   "scripts": {
     "lint": "eslint .",
-    "test": "jest",
-    "start": "babel-node index.js"
+    "test": "jest ./test",
+    "cp-package": "cp package.json ./dist",
+    "start": "babel-node ./dist/index.js",
+    "build": "rimraf ./dist && yarn tsc --project tsconfig.build.json"
   },
   "devDependencies": {
     "@babel/cli": "^7.1.0",
     "@babel/core": "^7.1.0",
     "@babel/node": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
+    "@babel/preset-typescript": "^7.17.12",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
     "babel-jest": "^27.5.1",
     "eslint": "^8.9.0",
     "eslint-config-prettier": "^3.1.0",
     "eslint-plugin-prettier": "^2.6.2",
-    "jest": "24.8.0",
+    "jest": "27.5.1",
     "prettier": "^1.14.3",
-    "regenerator-runtime": "^0.12.1"
+    "regenerator-runtime": "^0.12.1",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^28.0.3",
+    "typescript": "^4.6.4"
   }
 }
